Search Screen Style in dedicate file except for width adaptation
